benches: add benchmark for gauging complexity
These benchmarks wrap longer and longer texts and the goal is to get
numbers that can verify that the time grows as O(n) where n is the
size of the input text.

When running the benchmarks, the output looks like this on my machine:

  test lorem_100             ... bench:       2,085 ns/iter (+/- 11)
  test lorem_200             ... bench:       4,223 ns/iter (+/- 23)
  test lorem_400             ... bench:       8,331 ns/iter (+/- 36)
  test lorem_800             ... bench:      17,182 ns/iter (+/- 86)

That looks more or less like a linear growth.