benches: add hyphenation benchmarks
These tests show the performance hit of enabling automatic hyphenation
and also serve to verify the overall linear performance of the
library. On my machine, the numbers look like this:

  test hyphenation_lorem_100 ... bench:       4,527 ns/iter (+/- 61)
  test hyphenation_lorem_200 ... bench:       8,564 ns/iter (+/- 50)
  test hyphenation_lorem_400 ... bench:      17,913 ns/iter (+/- 79)
  test hyphenation_lorem_800 ... bench:      40,439 ns/iter (+/- 226)
  test lorem_100             ... bench:       2,085 ns/iter (+/- 11)
  test lorem_200             ... bench:       4,223 ns/iter (+/- 23)
  test lorem_400             ... bench:       8,331 ns/iter (+/- 36)
  test lorem_800             ... bench:      17,182 ns/iter (+/- 86)

Word wrapping a text with automatic hyphenation enabled takes roughly
twice as long as without, but we're still talking about operations
taking less than 100 microseconds, even for a long text.