WordSplitter: add trait for splitting words
Instead of putting the logic of how to split words into the Wrapper,
we now delegate this to a WordSplitter trait. There are two
implementations of this, corresponding to the old split_word method:

* HyphenSplitter: splits words on hyphens. This corresponds to having
  `corpus = None` before.

* Corpus: hyphenates words using language-specific hyphenation
  patterns. This corresponds to having `corpus = Some(...)` before.

This version is about 5% slower than the previous code. This could be
because of the extra indirection to the heap-allocated WordSplitter.