Wrapper: add builder methods
The new methods are named after the struct fields they modify. Using
them makes it easy to create and configure a Wrapper:

  let wrapper = Wrapper::new(78).break_words(false)
                                .initial_indent("* ")
                                .subsequent_indent("  ");

The methods consume self since that means one can create a wrapper
like above. If the methods had taken &mut self as suggested in

  https://github.com/brson/rust-api-guidelines#c-builder

the above would fail with

  error: borrowed value does not live long enough

since the Wrapper::new value is a temporary value that is dropped at
the end of the line.