Wrapper: make wrap_iter behave like wrap_iter_borrow
Before, we had three methods that would take a string and return an
iterator with wrapped lines:

* wrap_iter(&self, &str):        cloned self
* wrap_iter_borrow(&self, &str): borrowed self
* into_wrap_iter(self, &str):    consumed self

Here, the wrap_iter_borrow method is preferable to wrap_iter if the
Wrapper outlives the iterator.

I think that will be the usual case, and that we can do without the
method that cloned self -- if needed, the caller can get the
functionality back by combining clone and into_wrap_iter:

  wrapper.clone().into_wrap_iter(s)

We can then rename wrap_iter_borrow to just wrap_iter and thus make
the method with the short name do "the right thing" by default.

The new method names are thus:

* wrap_iter(&self, &str):        borrows self
* into_wrap_iter(self, &str):    consumes self

This makes the names a little more aligned with the standard library
where Vec::iter(&self) also borrows self while you iterate over the
collection.