tests: use check-versions to ensure README examples are updated
When I released Textwrap version 0.7.0 and 0.8.0, I forgot to update
the version numbers mentioned in the README in both cases. The
check-versions crate will help ensure that this doesn't happen again.

Since check-versions needs Rust 1.15.0 or later, this commit also
modifies our Travis config to remove the dev-dependency for the older
versions we test with. That way we can still test textwrap against
Rust 1.8.0, but still get the benefit of the check in newer versions.