cargo: make term_size optional
The term_size crate can be used to determine the current terminal
width and we included a trivial convenience wrapper around it.

However, term_size makes heavy use of unsafe Rust and this was a
concern to some users of textwrap. This commit makes the dependency a
non-default optional feature.

Fixes #101.