WrapIterImpl: take split_len into account when hyphenating
Before we did not take the width of the split into account when
hyphenating a word. This meant that all hyphenation was off by the
width of the whitespace (typically off by one when the split point was
a single space).

This has been broken since 4313d58 where I rewrote the splitting
algorithm to use Vec<Cow<str>> instead of Vec<String>. There is now a
test for this case.

Fixes #107.