travis: bump minimum version of Rust to 1.17.0
One of our dependencies bumped the minimum supported version of Rust
to 1.17.0 and even though we did not change anything, our builds
started failing.

We could try to find the troublesome dependency and pin it down, but
it seems that the general attitude towards version bumps of the
compiler is to simply support the latest two or three stable releases.
Rust 1.24 will be released later this week, so we still offer much
better support than most other libraries.