wrap_iter: simplify lifetimes in signature
Before, we declared the 'w lifetime for the Wrapper at the top of  the
impl block. This turns out to be unnecessary and we can move it to the
only method that uses it: wrap_iter.

We also declared the 'a lifetime as 'a: 'w. This constraint seems to
unnecessary, so it was removed.