travis: simplify build script
The --verbose flag seems to mostly make the output unreadable and has
thus been removed.

The caching might have made things faster, but it increases the risk
of us failing to spot problems when our dependencies are updated.

Putting the "cargo build" into the install script means that Travis
will hide the output by default. This lets us focus on the tests.