Explicitly mention when features are required
The documentation on https://docs.rs/textwrap/ is generated with the
--all-features flag, which means that it lists functions like
`termwidth`, which only exist when the `terminal_size` feature is
enabled.

This is confusing (see #165), especially when the feature name
changes, so that what is in `Cargo.toml` on GitHub doesn't match what
is documented on docs.rs.

Fixes #165.