Handle demo texts with more lines than what fits on the terminal
Before we would keep outputting lines, even if there are more lines
than what fits on the terminal. This resulted in garbage since the
bottom line would be overwritten repeatedly.

We now stop writing when we see that we have reached the bottom of the
terminal. In addition, we hide the cursor in this case to indicate
that the cursor is now “below” the bottom of the terminal.