Handle overflows in wrap_optimal_fit by divide-and-conquer
The `wrap_optimal_fit‘ algorithm computes the penalty for a gap as
`gap * gap`. If a fragment has a size near `usize::max_value()` and if
the line width is large, this computation can easily overflow.

When this happened, we would simply crash. Now, we instead do the
computations with checked arithmetic and detect the overflow. We then
proceed to wrap the first half of the fragments by themselves. If this
work, we then wrap the second half. This way, we might be able to wrap
everything without overflow.

Should there be a single fragment which causes the overflow by itself,
this fragment is put on a line by itself.

When wrapping part of the fragments, we might of course end up with a
partial last line. To fix this, we pop the last line and re-wrap the
fragments that were put onto this line. This helps remove some of the
“seams” that would otherwise occur. However, this is not perfect if
there are many over-sized fragments since they can still cause
half-empty lines to appear in the output.

Fixes #247.