Include a dependency graph in the generated documentation
This used cargo-depgraph to generate a dependency graph which we
convert to a SVG image. The image is committed to the repository under
a versioned filename. This allows us to link to the correct image in
the documentation.

The dependency graph is updated when new tags are pushed — this should
ensure that the images are kept up to date.