Extend display_width to handle emojis when unicode-width is disabled
The unicode-width Cargo feature is used by default to measure the
width of the characters. Without it, we used to fall back to a simple
model where each character is has a width of one column.

However, this is wrong for emojis such as “😂”, which normally take up
two columns when printed in a terminal. We now handle such characters
as well — we simply assume that all characters above U+1100 is a
double-width character.

This turns out to be correct for emojis, but it is wrong for many
non-emoji characters. To handle those, the unicode-width feature
should simply be enabled.