Add `WordSeparator` trait allow customizing how words are found
The new trait is responsible for turning a line of text into an
iterator of `Word`s. The first implementation is `AsciiSpace`, which
simply splits the input text on `' '`. This matches the current
behavior.

The new trait shows up as an extra generic parameter on `Options`.
This is quite invasive in the case of dynamic dispatch: when you could
write

```rust
let options: Options = Options::with_splitter(80, NoHyphenation);
```

before to create a `Options<Box<dyn WordSplitter>>`, you now need to
cast two types to trait objects to achive the same thing:

```rust
let mut options = Options::new(80)
    .splitter(Box::new(HyphenSplitter) as Box<dyn WordSplitter>)
    .word_separator(Box::new(AsciiSpace) as Box<dyn WordSeparator>);
```