Replace the `core::WrapAlgorithm` enum with a `WrapAlgorithm` trait
This introduces a `textwrap::wrap_algorithms` module where the
`wrap_first_fit` and `wrap_optimal_fit` functions now live.

A new `WrapAlgorithm` trait replaces the old enum. This allows for
arbitrary wrapping algorithms to be plugged into the library. We
should also be able to change the return type from `Vec<&[Word]>` to
`Iterator<Item = Word>` and thus implement streaming wrapping for the
first-fit in the future.

Fixes #325.