Cache cargo-fuzz between invocations
This should save us about two minutes when fuzzing since we can reuse
the cargo-fuzz binary between runs.

While nightly Rust is required to do fuzzing, the cargo-fuzz binary is
not linked directly to the precise version of the nightly compiler
being used — I’ve been using a 0.8.0 release of cargo-fuzz for months
and it seems happy enough. We therefore link the cache with the latest
version number of cargo-fuzz itself, thus ensuring that we always use
the latest version.