Cache fuzz testing corpus between runs
This should give us an accumulating cache with the fuzz testing
corpus. Each run will see a cache miss, and fallback to downloading
the most recent cache. After the run, a new cache is uploaded.

The caches are small (3-5 MB) and deleted after a week, so the total
amount of data used is in the order of 30-40 MB.