Use floating point math to compute line penalties
This changes the line penalty computation from a simple

    (target_width - line_width)**2

to a scaled penalty which ranges from 0.0 (a perfect fit) to a
variable `max_line_penalty` (an infinitely bad fit):

    ((target_width - line_width) / target_width)**2 * max_line_penalty

Scaling the penalties in this sense avoids the overflows that could
happen in the past when `target_width` was large, so that the gap at
the end of line could be come very large. Now the line cost always
goes from `0` to `max_line_penalty`, regardless of the actual target
width.

The cost still increases in a quadratic fashion. The
`max_line_penalty` is set to 10,000.0. This makes the new cost
computations identical to the past computations if the target width is
100 columns wide.