Add demo program to help compute binary sizes
The demo program can be compiled with and without Textwrap. This
allows us to compare the size of a quick-and-dirty implementation with
the implementation in Textwrap.

With Rust 1.54 (current stable), I see these results:

| Configuration                            |  Binary Size |    Delta |
| :---                                     |         ---: |     ---: |
| quick-and-dirty implementation           |       252 KB |     — KB |
| textwrap without default features        |       268 KB |    16 KB |
| textwrap with smawk                      |       284 KB |    32 KB |
| textwrap with unicode-width              |       276 KB |    24 KB |
| textwrap with unicode-linebreak          |       362 KB |   110 KB |

The sizes are the stripped sizes and the binary is compiled in release
mode with this profile:

```
[profile.release]
lto = true
codegen-units = 1
```