Use precise dependency versions in Cargo.toml
Following the advice from [1], this PR updates Cargo.toml to use
precise version numbers for all dependencies. The latest versions at
the time of writing are used.

It turns out that we made precisely the mistake mentioned in the post:

```
% cargo -Z minimal-versions update
% cargo test
```

failed on nightly because of the dependency on smawk 0.3: we need
0.3.1 after #421.

[1]: https://users.rust-lang.org/t/psa-please-specify-precise-dependency-versions-in-cargo-toml/71277