Remove `Options::with_word_splitter`
This constructor was a left over from back when `Options` had a single
`WordSplitter` generic type parameter. Now that we have several
generic type parameters, having this constructor make one parameter
different from the others (for no good reason).

The `with_word_splitter` method was const, so we now lose the
ability to construct `Options` with a non-default `WordSplitter` in a
const context. If this is important to you, please get in contact so
we can look at making the builder methods const instead.