Revert "Put single-expression functions on a single line"
This reverts commit 4ce32d6d878fe9c500e331b3553c9f6d7ae2d145.

Changing the default formatting using an unstable `rustfmt` feature
was a bad idea: it means that people who have auto-formatting turned
on (like me) will see the formatting flip-flop when switching between
stable and nightly.

The problem is that the unstable `fn_single_line` option only takes
effect on nightly, so when `rustfmt` runs on stable, it will add lots
of spurious changes.

The `imports_granularity` option does not have this problem: `rustfmt`
will not modify imports by default and so nothing happens when it is
run on stable.