Fix crash in `unfill` when string ends with newlines
Test that we don't crash on a `\r` following a string of `\n`. The
problem was that we removed both kinds of characters in one code path,
but not in the other. We now use the same `text` in both places and
this fixes the issue.

The fix changed the handling of newlines slightly: we now normalize
the newlines to a single `\n` or `\r\n` depending on what we detect in
the input. The documentation was updated to explicitly say that
passing in multiple paragraphs result in unspecified behavior.