Add fast-path to `fill` when input is shorter than wrap width
This adds a conservative check to `fill` which will avoid wrapping the
input when it is a single line which is shorter than the wrap width.
In that case, wrapping the input is now 10-25 times faster than
before.

The input length is measured in bytes, which means that we can
overestimate the length compared `display_width`. This is because
multi-byte sequences (such as combining diacritics) can be displayed
as a single character. So we will end up taking the slow path for some
strings, but the vast majority of short strings will take the fast
path.

The correctness of the fast path is checked with a new fuzz test.