Limit inputs to prevent OSS-Fuzz timeouts
I downloaded the failing test cases from OSS-Fuzz and ran the fuzzers
on each of them: they all finished in ~35 seconds on my laptop. The
timeout on OSS-Fuzz is 25 seconds, which explains the timeouts.

Limiting the input length to 100k bytes should help here and still
allow us to test all reasonable inputs.