Fix Clippy warnings
Running Clippy in CI seems like a recipe for problems: the stable
Clippy will change its opinion on a lint (as happened with
`clippy::uninlined-format-args`) and the build will suddenly fail.

It’s still nice to run it once in a while, but I’ll do that by hand
for now.