Test with stable and nightly, build with 1.56
The one-month old fix in #515 has already caused problems because a
test dependency has bumped its minimum supported Rust version (MSRV).

It turns out that Textwrap builds fine on old versions of Rust. It
builds all the way back to Rust 1.56 which introduced the 2021 edition
we use. To better reflect this, we now _test_ with modern versions of
Rust and we build with the old version.