Add example showing how one can use owned options
The `textwrap::Options<'a>` type is a little annoying: it insists on
borrowing all its fields from somewhere else. In particular, the
indentation fields are of type `&'a str`. This means that it’s not
really possible for a program to store an `Options<'a>` and use it
later if the fields are anything but string literals (of type `&str`).

However, what does work is to create the `Options<'a>` on demand.
Since all functions that need the options take an argument of type
`impl Options<'a>`, you can easily convert your own type as needed.
This can be as cheap as possible, i.e., converting `String` fields to
`&str`.

From the discussion in #592.